# contracts/backend/openapi.yaml
openapi: 3.1.0
info:
  title: DIP3 Java Analysis Backend
  version: 2.0.0
paths:
...
  /api/admin/ontology/overview:
    get:
      summary: Return Java-owned ontology governance overview
      security: [{ cookieAuth: [] }]
      responses:
        '200': { description: Governance overview, content: { application/json: { schema: { $ref: './schemas/governance-overview.schema.json' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/definitions:
    get:
      summary: Return all governed definition categories for one version
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: versionId, in: query, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Version definitions, content: { application/json: { schema: { $ref: './schemas/governance-definitions.schema.json' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/versions:
    get:
      summary: List ontology versions available to governance users
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        '200':
          description: Version list
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [items, capabilities]
                properties:
                  items: { type: array, items: { $ref: './schemas/governance-shared.schema.json#/$defs/version' } }
                  capabilities: { $ref: './schemas/governance-shared.schema.json#/$defs/capabilities' }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/change-requests:
    get:
      summary: List governed change requests
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: status, in: query, required: false, schema: { type: string } }
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        '200': { description: Change request list, content: { application/json: { schema: { $ref: './schemas/governance-change-request-list.schema.json' } } } }
    post:
      summary: Create a draft change request in Java
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [ontologyVersionId, targetObjectType, targetObjectKey, changeType, title, impactScope, compatibilityType]
              properties:
                ontologyVersionId: { type: string, minLength: 1, maxLength: 200 }
                targetObjectType: { enum: [entity_definition, metric_definition, metric_variant, factor_definition, causality_edge, plan_step_template, tool_capability_binding, time_semantic, evidence_type_definition] }
                targetObjectKey: { type: string, minLength: 1, maxLength: 200 }
                changeType: { enum: [create, update, deprecate, retire] }
                title: { type: string, minLength: 1, maxLength: 200 }
                description: { type: [string, 'null'], maxLength: 2000 }
                beforeSummary: { type: [object, 'null'] }
                afterSummary: { type: [object, 'null'] }
                impactScope: { type: array, items: { type: string, minLength: 1, maxLength: 200 } }
                compatibilityType: { enum: [backward_compatible, breaking] }
                compatibilityNote: { type: [string, 'null'], maxLength: 1000 }
      responses:
        '200': { description: Created change request, content: { application/json: { schema: { $ref: './schemas/governance-shared.schema.json#/$defs/changeRequest' } } } }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/change-requests/{id}:
    get:
      summary: Return a governed change request and approval history
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Change request detail, content: { application/json: { schema: { $ref: './schemas/governance-change-request-detail.schema.json' } } } }
        '404': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/change-requests/{id}/submit:
    post:
      summary: Submit one draft change request for review
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, minLength: 1 } }
      responses:
        '200': { description: Submitted change request, content: { application/json: { schema: { $ref: './schemas/governance-shared.schema.json#/$defs/changeRequest' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/change-requests/{id}/review:
    post:
      summary: Approve or reject one submitted change request
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, minLength: 1 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [decision]
              properties:
                decision: { enum: [approved, rejected] }
                comment: { type: [string, 'null'], maxLength: 1000 }
      responses:
        '200': { description: Review result }
        '400': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/publishes:
    get:
      summary: List Java-owned ontology publish records
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: limit, in: query, required: false, schema: { type: integer, minimum: 1, maximum: 200 } }
      responses:
        '200': { description: Publish history, content: { application/json: { schema: { $ref: './schemas/governance-publish-history.schema.json' } } } }
        '401': { $ref: '#/components/responses/Error' }
        '403': { $ref: '#/components/responses/Error' }
  /api/admin/ontology/versions/{id}/publish:
    post:
      summary: Atomically publish an approved ontology version
      security: [{ cookieAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, minLength: 1 } }
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              properties:
                publishNote: { type: [string, 'null'], maxLength: 1000 }
      responses:
        '200': { description: Publish record, content: { application/json: { schema: { $ref: './schemas/governance-shared.schema.json#/$defs/publish' } } } }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
        '409': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    cookieAuth: { type: apiKey, in: cookie, name: dip3_session }
    graphSyncOpsSecret: { type: apiKey, in: header, name: X-Graph-Sync-Ops-Secret }
  responses:
    Error:
      description: Explicit failure
      content:
        application/json: { schema: { $ref: './schemas/error.schema.json' } }
...